Added feature to let user enter a goal in hours for the week. A goal is represented by a Goal entity that is inserted into database. Query database for a goal within the current week. If a goal exists goalBarChart in WeekViewFragment.kt is updated with a limit line set to this goals hours. Users total hours for the last seven days are displayed on goalBarChart.
WeekViewFragment.kt updates weekBarChart with users hours from last seven days. goalBarChart is updated when user sets a goal for the week. Hours are updated and moved towards the goal in the bar chart as user adds new study sessions.

Added functions to Repository.kt for adding a weekly goal and monthly goal.