Added Custom Calendar to MonthDetailFragment.kt. for loop sets views text according to the the month selected's first day of the month.
MonthDetailViewModel.kt uses monthlyHours live data to display the days for which a user has hours in the custom calendar view.

MonthDetailFragment.kt xml file has text views for days which are ordered in xml to match indices 0 - 41. Text views for hours which are indexed from 42 - 78

For loops to set days starts with the first day of the month selected. So if the first day of the month is 3 then day text view 3 is set to 1

For loop to set hours text view would start on 3 + 42 + 3 = hours text view  index 48